Move `FuturesStream` to a separate file
