+++ src/protocol/libp2p/kademlia/executor.rs
@@ -18,15 +18,18 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-use crate::{protocol::libp2p::kademlia::query::QueryId, substream::Substream, PeerId};
+use crate::{
+    protocol::libp2p::kademlia::{futures_stream::FuturesStream, query::QueryId},
+    substream::Substream,
+    PeerId,
+};
 
 use bytes::{Bytes, BytesMut};
-use futures::{future::BoxFuture, stream::FuturesUnordered, Stream, StreamExt};
+use futures::{future::BoxFuture, Stream, StreamExt};
 
 use std::{
-    future::Future,
     pin::Pin,
-    task::{Context, Poll, Waker},
+    task::{Context, Poll},
     time::Duration,
 };
 
@@ -69,53 +72,6 @@ pub struct QueryContext {
 
     /// Query result.
     pub result: QueryResult,
-}
-
-/// Wrapper around [`FuturesUnordered`] that wakes a task up automatically.
-#[derive(Default)]
-pub struct FuturesStream<F> {
-    futures: FuturesUnordered<F>,
-    waker: Option<Waker>,
-}
-
-impl<F> FuturesStream<F> {
-    /// Create new [`FuturesStream`].
-    pub fn new() -> Self {
-        Self {
-            futures: FuturesUnordered::new(),
-            waker: None,
-        }
-    }
-
-    /// Push a future for processing.
-    pub fn push(&mut self, future: F) {
-        self.futures.push(future);
-
-        if let Some(waker) = self.waker.take() {
-            waker.wake();
-        }
-    }
-}
-
-impl<F: Future> Stream for FuturesStream<F> {
-    type Item = <F as Future>::Output;
-
-    fn poll_next(mut self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
-        let Poll::Ready(Some(result)) = self.futures.poll_next_unpin(cx) else {
-            // We must save the current waker to wake up the task when new futures are inserted.
-            //
-            // Otherwise, simply returning `Poll::Pending` here would cause the task to never be
-            // woken up again.
-            //
-            // We were previously relying on some other task from the `loop tokio::select!` to
-            // finish.
-            self.waker = Some(cx.waker().clone());
-
-            return Poll::Pending;
-        };
-
-        Poll::Ready(Some(result))
-    }
 }
 
 /// Query executor.
