+++ src/protocol/libp2p/kademlia/mod.rs
@@ -68,6 +68,7 @@ const PARALLELISM_FACTOR: usize = 3;
 mod bucket;
 mod config;
 mod executor;
+mod futures_stream;
 mod handle;
 mod message;
 mod query;
